backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
celery[redis]==5.4.0
redis==5.1.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
stripe==10.0.0
pyotp==2.9.0
qrcode==7.4.2
boto3==1.35.0
openpyxl==3.1.5
phonenumbers==8.13.0
python-dotenv==1.0.1
email-validator==2.2.0
jinja2==3.1.4
aiofiles==24.1.0
websockets==13.0
tenacity==9.0.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx[http2]==0.27.0
